Don't run nvm.sh unless necessary
Running this in .zshrc was causing a >5 second slowdown every time a new
shell was spawned.

This solution was taken from
https://github.com/nvm-sh/nvm/issues/1277#issuecomment-536218082
and just links the correct version of node statically, which is really
what nvm should be doing anyway, meaning no startup time.

This requires running `nvm alias default <version number>` to ensure
that a version number is present in the nvm default file.

nvm is still available by running nvm.sh the first time you call it